+++ src/scad_tarmi/dfm.clj
@@ -29,11 +29,16 @@
   in this manner is typically difficult but saves on transformations in the
   OpenSCAD output, thus improving rendering performance.
 
-  Called with more arguments, the closure implements scad-tarmi’s maybe/scale."
+  Called with two arguments, a number and an option map, the closure again
+  adjusts the stated dimensional input, and can do so even for a positive
+  error, when non-default {:negative false} is part of the option map.
+
+  Called with more arguments, the closure implements scad-tarmi’s maybe/scale,
+  returning scad-clj code, not a number."
   ([] (error-fn (- 0.5)))
   ([error] (error-fn (/ error -2) error))
   ([positive-error negative-error]
-   (fn
+   (fn compensator
      ([nominal] (- nominal negative-error))
      ([nominal options & block]
       {:pre [(number? nominal)
@@ -42,7 +47,12 @@
              :or {negative true, x true, y true, z false}} options
             n (ratio nominal (if negative negative-error positive-error))
             factors (vec (map #(if % n 1) [x y z]))]
-        (apply (partial maybe/scale factors) block))))))
+        (if (nil? block)
+          ;; No shape was passed. Return a number.
+          (if negative
+            (compensator nominal)  ; Recurse to the simpler form.
+            (- nominal positive-error))
+          (apply (partial maybe/scale factors) block)))))))
 
 (def none
   "An error function that leaves no trace in OpenSCAD code."